Uprava ted fiči! Všechno je ted v češtině! Potvrzovani u smazani přidano! Při a po upravě se správně načte čas!

+++ CSAA V2/CSAA/CSAA/src/main/webapp/reactjs/src/components/ReservationList.js
@@ -30,15 +30,16 @@ export default class ReservationList extends Component {
 
 
     deleteReservation = (reservationId) =>{
-        axios.delete("http://localhost:8080/reservations/"+reservationId)
-            .then(response =>{
-                if(response.data != null){
-                    alert("Smazáno!");
-                    this.setState({
-                        reservations: this.state.reservations.filter(reservation => reservation.id !== reservationId)
-                    });
-                }
-            })
+        if(window.confirm("Jste si jistý smazáním této rezervace?")){
+            axios.delete("http://localhost:8080/reservations/"+reservationId)
+                .then(response =>{
+                    if(response.data != null){
+                        this.setState({
+                            reservations: this.state.reservations.filter(reservation => reservation.id !== reservationId)
+                        });
+                    }
+                })
+        }
     };
 
     render(){
